refactor(openai): move data_residency helper under llms/openai
